+++ Tchaso/src/app/Specialites/specialite/specialite.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
+import { ServiceutilisateurService } from 'src/app/Utilisateurs/service/serviceutilisateur.service';
 import Swal from 'sweetalert2';
 import { ServicespecialiteService } from '../servicespecialite.service';
 
@@ -11,7 +12,7 @@ import { ServicespecialiteService } from '../servicespecialite.service';
 export class SpecialiteComponent implements OnInit {
   searchText= '';
   listeSpecialites: any = [];
-  constructor(public service: ServicespecialiteService, public router: Router) { }
+  constructor(public service: ServicespecialiteService, public serviceUser: ServiceutilisateurService, public router: Router) { }
 
   ngOnInit(): void {
     this.getAllSpecialite();    
@@ -31,23 +32,58 @@ export class SpecialiteComponent implements OnInit {
     this.service.deleteSpecialite(id).subscribe((data: any)=>{})
   }
 
+  deleteSpecialiteTravailleur(id: any){
+    this.service.getTravailleurBySpecialite(id).subscribe((datas: any)=>{
+      if (datas.length > 0){
+        this.service.deleteSpecialite(id).subscribe((data: any)=>{
+          for(let i=0; i<datas.length; i++){
+            this.serviceUser.deleteTravailleur(datas[i].id).subscribe((spe:any)=>{});
+          }
+        })
+       
+      }
+    })
+  }
 
   alertConfirmation(id: any) {
-    Swal.fire({
-      title: 'ATTENTION',
-      text: 'Êtes vous sûr de vouloir supprimer cette spécialité ?',
-      icon: 'warning',
-      showCancelButton: true,
-      confirmButtonText: 'Supprimer',
-      cancelButtonText: 'Annuler',
-    }).then((result) => {
-      if (result.value) {
-        this.deleteSpecialite(id);
-        Swal.fire('Suppression!', 'Spécialité supprimée avec succès.', 'success');
-        window.location.reload();
-        this.router.navigateByUrl('specialites', {skipLocationChange: true}).then(()=>
-        this.router.navigate(['specialites']));
+    this.service.getTravailleurBySpecialite(id).subscribe((data: any)=>{
+      if (data.length > 0){
+        Swal.fire({
+          title: 'ATTENTION',
+          text: 'Des travailleurs sont rattachés à cette spécialité, êtes vous sûr de vouloir supprimer cette spécialité ?',
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonText: 'Supprimer',
+          cancelButtonText: 'Annuler',
+        }).then((result) => {
+          if (result.value) {
+            this.deleteSpecialiteTravailleur(id);
+            Swal.fire('Suppression!', 'Spécialité supprimée avec succès.', 'success');
+            window.location.reload();
+            this.router.navigateByUrl('specialites', {skipLocationChange: true}).then(()=>
+            this.router.navigate(['specialites']));
+          }
+        });
       }
-    });
+      if (data.length == 0){
+        Swal.fire({
+          title: 'ATTENTION',
+          text: 'Êtes vous sûr de vouloir supprimer cette spécialité ?',
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonText: 'Supprimer',
+          cancelButtonText: 'Annuler',
+        }).then((result) => {
+          if (result.value) {
+            this.deleteSpecialite(id);
+            Swal.fire('Suppression!', 'Spécialité supprimée avec succès.', 'success');
+            window.location.reload();
+            this.router.navigateByUrl('specialites', {skipLocationChange: true}).then(()=>
+            this.router.navigate(['specialites']));
+          }
+        });
+      }
+    })
+
   }
 }
